requirements.txt
anyio
asgiref
certifi
charset-normalizer
Django
djangorestframework
djangorestframework-simplejwt
h11
httpcore
idna
psycopg2-binary
PyJWT
python-dotenv
requests
sniffio
sqlparse
tzdata
urllib3
Pillow